Execution context awareness (#3)
Analyzes the stack to determine if the implant is awaken from within a JUnit test, build tool (Maven or Gradle) or if it runs standalone. This can be used to do different things inside the payload.

Also improves the target hostname check to use a regular expression instead of a simple string quality check.